requirements.txt
matplotlib
numpy
Pillow
torch
ray
pandas
torchvision
ray[tune]